# apis/family/openapi.yaml
openapi: 3.1.0
info:
  title: Family 360 - Family API
  version: 0.2.0
  description: >
    API for managing families and their direct relationships. 
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Family API major version 1
...
security:
  - bearerAuth: []
paths:
  /families:
    get:
      tags:
        - Families
      summary: List families available to the current user
      operationId: listFamilies
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Families page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyPage'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags:
        - Families
      summary: Create a family
      operationId: createFamily
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyInput'
      responses:
        '201':
          description: Family created
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /families/{familyId}:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    get:
      tags:
        - Families
      summary: Get a family
      operationId: getFamily
      responses:
        '200':
          description: Family found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
    patch:
      tags:
        - Families
      summary: Partially update a family
      operationId: patchFamily
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyPatch'
      responses:
        '200':
          description: Family updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags:
        - Families
      summary: Disable a family
      operationId: disableFamily
      responses:
        '204':
          description: Family disabled
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
...
  /families/{familyId}/members:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    get:
      tags:
        - Family Members
      summary: List family members
      operationId: listFamilyMembers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Family members page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyMemberPage'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags:
        - Family Members
      summary: Create a family member
      operationId: createFamilyMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyMemberInput'
      responses:
        '201':
          description: Family member created
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyMember'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /families/{familyId}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/MemberId'
    get:
      tags:
        - Family Members
      summary: Get a family member
      operationId: getFamilyMember
      responses:
        '200':
          description: Family member found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyMember'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
    patch:
      tags:
        - Family Members
      summary: Partially update a family member
      operationId: patchFamilyMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyMemberPatch'
      responses:
        '200':
          description: Family member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyMember'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags:
        - Family Members
      summary: Disable a family member
      operationId: disableFamilyMember
      responses:
        '204':
          description: Family member disabled
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        default:
          $ref: '#/components/responses/ProblemResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    Location:
      description: URI of the created resource
      schema:
        type: string
        format: uri
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      description: Maximum number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: Sort expression, such as name,asc
      schema:
        type: string
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ProblemResponse:
      description: Error response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnauthorizedProblem:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ForbiddenProblem:
      description: The current user cannot perform the operation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFoundProblem:
      description: The requested family resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationProblem:
      description: The request violates input or domain validation rules
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    PageMetadata:
      type: object
      required:
        - page
        - limit
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
    ProblemError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
        pointer:
          type: string
    FamilyBase:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: EUR
    FamilyInput:
      allOf:
        - $ref: '#/components/schemas/FamilyBase'
        - type: object
          required:
            - name
            - currency
    FamilyPatch:
      allOf:
        - $ref: '#/components/schemas/FamilyBase'
    Family:
      allOf:
        - $ref: '#/components/schemas/FamilyBase'
        - type: object
          required:
            - id
            - name
            - currency
            - createdAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            disabledAt:
              type:
                - string
                - 'null'
              format: date-time
    FamilyPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Family'
...
    FamilyMemberBase:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        memberType:
          $ref: '#/components/schemas/FamilyMemberType'
        birthDate:
          type:
            - string
            - 'null'
          format: date
    FamilyMemberInput:
      allOf:
        - $ref: '#/components/schemas/FamilyMemberBase'
        - type: object
          required:
            - name
            - memberType
    FamilyMemberPatch:
      allOf:
        - $ref: '#/components/schemas/FamilyMemberBase'
    FamilyMember:
      allOf:
        - $ref: '#/components/schemas/FamilyMemberBase'
        - type: object
          required:
            - id
            - familyId
            - name
            - memberType
            - createdAt
          properties:
            id:
              type: string
              format: uuid
            familyId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            disabledAt:
              type:
                - string
                - 'null'
              format: date-time
    FamilyMemberPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/FamilyMember'
...
    FamilyMemberType:
      type: string
      enum:
        - ADULT
        - CHILD
        - DEPENDENT
        - OTHER
...